+++ mods/include/mods/sign/sign.h
@@ -1,5 +1,0 @@
-#pragma once
-
-extern "C" {
-void sign_key_press(char key);
-}