+++ mods/src/compat/sdl.cpp
@@ -1,43 +1,85 @@
 #include <SDL/SDL.h>
-
 #include <media-layer/core.h>
-#include <libreborn/patch.h>
+
+#include <libreborn/util.h>
+#include <libreborn/config.h>
+
+#include <symbols/minecraft.h>
+
+#include <mods/compat/compat.h>
+#include <mods/input/input.h>
+#include <mods/feature/feature.h>
+
 #include "compat-internal.h"
 
-// SDL Stub
-static void *SDL_SetVideoMode_injection(__attribute__((unused)) int width, __attribute__((unused)) int height, __attribute__((unused)) int bpp, __attribute__((unused)) uint32_t flags) {
-    // Return Value Is Only Used For A NULL-Check
-    return (void *) 1;
+// Custom Title
+HOOK(media_SDL_WM_SetCaption, void, (__attribute__((unused)) const char *title, const char *icon)) {
+    real_media_SDL_WM_SetCaption()(MCPI_APP_TITLE, icon);
 }
 
-// Window Information
-static void x11_nop() {
-    // NOP
-}
-static int SDL_GetWMInfo_injection(SDL_SysWMinfo *info) {
-    // Return Fake Lock Functions Since XLib Isn't Directly Used
-    SDL_SysWMinfo ret;
-    ret.info.x11.lock_func = x11_nop;
-    ret.info.x11.unlock_func = x11_nop;
-    ret.info.x11.display = nullptr;
-    ret.info.x11.window = 0;
-    ret.info.x11.wmwindow = ret.info.x11.window;
-    *info = ret;
-    return 1;
+// Mouse Cursor Is Always Invisible In Vanilla MCPI
+// Because In Vanilla MCPI, The GPU Overlay Covered The Normal Mouse Cursor
+HOOK(media_SDL_ShowCursor, int, (int toggle)) {
+    return real_media_SDL_ShowCursor()(toggle == SDL_QUERY ? SDL_QUERY : SDL_DISABLE);
 }
 
-// Quit
-static void SDL_Quit_injection() {
-    // Cleanup Media Layer
-    media_cleanup();
-    // Exit
-    INFO("Stopped");
+// Intercept SDL Events
+static bool enable_text_events;
+HOOK(media_SDL_PollEvent, int, (SDL_Event *event)) {
+    // In Server Mode, Exit Requests Are Handled In src/server/server.cpp
+    // Check If Exit Is Requested
+    if (!reborn_is_server() && compat_check_exit_requested()) {
+        // Send SDL_QUIT
+        SDL_Event new_event;
+        new_event.type = SDL_QUIT;
+        media_SDL_PushEvent(&new_event);
+    }
+
+    // Poll Events
+    int ret = real_media_SDL_PollEvent()(event);
+
+    // Handle Events
+    if (ret == 1 && event != nullptr) {
+        bool handled = false;
+        switch (event->type) {
+            case SDL_KEYDOWN:
+            case SDL_KEYUP: {
+                // Track Control Key
+                bool is_ctrl = (event->key.keysym.mod & KMOD_CTRL) != 0;
+                if (event->type == SDL_KEYUP) {
+                    is_ctrl = false;
+                }
+                input_set_is_ctrl(is_ctrl);
+                break;
+            }
+            case SDL_MOUSEBUTTONDOWN:
+            case SDL_MOUSEBUTTONUP: {
+                // Track Right-Click State
+                if (event->button.button == SDL_BUTTON_RIGHT) {
+                    input_set_is_right_click(event->button.state != SDL_RELEASED);
+                }
+                break;
+            }
+            case SDL_USEREVENT: {
+                // SDL_UserEvent Is Never Used In MCPI, So It Is Repurposed For Character Events
+                if (event->user.code == USER_EVENT_CHARACTER) {
+                    if (enable_text_events) {
+                        Keyboard::_inputText.push_back(char(event->user.data1));
+                    }
+                    handled = true;
+                }
+                break;
+            }
+        }
+        if (handled) {
+            // Event Was Handled
+            return media_SDL_PollEvent(event);
+        }
+    }
+    return ret;
 }
 
-// Patch SDL Calls
-void _patch_sdl_calls() {
-    // Disable SDL Calls
-    overwrite_call((void *) 0xe020, (void *) SDL_SetVideoMode_injection);
-    overwrite_call((void *) 0x13284, (void *) SDL_GetWMInfo_injection);
-    overwrite_call((void *) 0x12410, (void *) SDL_Quit_injection);
+// Init
+void _init_compat_sdl() {
+    enable_text_events = feature_has("Enable Text Input", server_disabled);
 }